services/server/requirements.txt
flasgger==0.9.2
Flask==1.0.2
Flask-Cors==3.0.4
flask-migrate==2.2.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
psycopg2-binary==2.7.7
